Analytics: Provide API for clearing persistent contexts
This will be necessary when users sign out
