+++ PocketKit/Sources/Analytics/Context/APIUserContext.swift
@@ -2,6 +2,9 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
+import Foundation
+
+
 public struct APIUserContext: Context {
     public static var schema = "iglu:com.pocket/api_user/jsonschema/1-0-1"
     
@@ -11,6 +14,20 @@ public struct APIUserContext: Context {
     public init(id: UInt, clientVersion: String) {
         self.id = id
         self.clientVersion = clientVersion
+    }
+
+    public init(consumerKey: String) {
+        let components = consumerKey.components(separatedBy: "-")
+        let id: UInt
+        if let identifier = components.first, let apiID = UInt(identifier) {
+            id = apiID
+        } else {
+            id = 1
+        }
+
+        let clientVersion = Bundle.main.object(forInfoDictionaryKey: "CFBundleVersion") as! String
+
+        self.init(id: id, clientVersion: clientVersion)
     }
 }
